structs: change type of ply to int32_t
from uint32_t

fix warnings for comparison between signed and unsigned types
remove need for type casts to int32_t

+++ src/search.c
@@ -53,7 +53,7 @@ int32_t negamax(uint32_t depth, int32_t alpha, int32_t beta,
 
    int32_t best = -INFSCORE + state.ply;
    if (!principal && (best = probe_hash(depth, &alpha, &beta, &store))
-         != (int32_t)(-INFSCORE + state.ply)) { return best; }
+         != -INFSCORE + state.ply) { return best; }
 
    if (!depth) { return quiescence(alpha, beta); }
 
@@ -81,7 +81,7 @@ int32_t negamax(uint32_t depth, int32_t alpha, int32_t beta,
       __builtin_prefetch(&ttable[state.hash & (HASHMASK ^ 0x3)], 1, 3);
       tree_node_entry(alpha, beta);
       int32_t score;
-      if (best == (int32_t)(-INFSCORE + state.ply)) {
+      if (best == -INFSCORE + state.ply) {
          score = -negamax(depth - 1, -beta, -alpha, principal);
       } else {
          score = -negamax(depth - 1, -alpha - 1, -alpha, 0);
